LMS-2857 moodle-logstore_xapi: update moveback test after we introduced type handling in the tasks.
